format code

+++ contracts/examples/Xyn.sol
@@ -11,16 +11,15 @@ interface IFactory {
 /// @title A LayerZero OmnichainFungibleToken example of BasedOFT
 /// @notice Use this contract only on the BASE CHAIN. It locks tokens on source, on outgoing send(), and unlocks tokens when receiving from other chains.
 contract TestToken is BasedOFT {
-   
     bool private tradingActivated = false;
-    uint256 private blockForPenaltyEnd;
+    uint private blockForPenaltyEnd;
     address public lpPairLaunch = address(0);
     address private tokensaleContract = address(0);
     address public FACTORY;
     address public WETH;
     address public ROUTER;
-    mapping (address => bool) private restrictedWallets;
-    mapping (address => bool) private automatedMarketMakerPairs;
+    mapping(address => bool) private restrictedWallets;
+    mapping(address => bool) private automatedMarketMakerPairs;
 
     constructor(address _layerZeroEndpoint, uint _initialSupply, address _FACTORY, address _WETH, address _ROUTER) BasedOFT("AIT", "AIT", _layerZeroEndpoint) {
         _mint(_msgSender(), _initialSupply);
@@ -31,41 +30,36 @@ contract TestToken is BasedOFT {
         lpPairLaunch = IFactory(FACTORY).createPair(address(this), WETH);
     }
 
-
-    function _transfer(
-        address from,
-        address to,
-        uint256 amount
-    ) internal override {
+    function _transfer(address from, address to, uint amount) internal override {
         require(from != address(0), "ERC20: transfer from the zero address");
 
-        if(!tradingActivated){
+        if (!tradingActivated) {
             require(_msgSender() == owner() || _msgSender() == tokensaleContract || _msgSender() == ROUTER, "Trading is not active.");
         }
 
-        if(earlyBuyPenaltyInEffect() && from == lpPairLaunch && to != lpPairLaunch) {
-            if(!restrictedWallets[to]) {
+        if (earlyBuyPenaltyInEffect() && from == lpPairLaunch && to != lpPairLaunch) {
+            if (!restrictedWallets[to]) {
                 restrictedWallets[to] = true;
-             }
+            }
         }
 
-        if(!earlyBuyPenaltyInEffect() && tradingActivated){
+        if (!earlyBuyPenaltyInEffect() && tradingActivated) {
             require(!restrictedWallets[from] || to == owner() || to == address(0xdead), "Bots cannot transfer tokens in or out except to owner or dead address.");
         }
 
         super._transfer(from, to, amount);
     }
 
-    function earlyBuyPenaltyInEffect() private view returns (bool){
+    function earlyBuyPenaltyInEffect() private view returns (bool) {
         return block.number < blockForPenaltyEnd;
     }
 
-    function enableTrading(uint256 blocksForPenalty) external onlyOwner {
+    function enableTrading(uint blocksForPenalty) external onlyOwner {
         require(lpPairLaunch != address(0), "Lp pair not set");
         require(!tradingActivated, "Trading is already active, cannot relaunch.");
         require(blocksForPenalty < 3, "Cannot make penalty blocks more than 2");
         tradingActivated = true;
-        uint256 tradingActivatedBlock = block.number;
+        uint tradingActivatedBlock = block.number;
         blockForPenaltyEnd = tradingActivatedBlock + blocksForPenalty;
     }
 
@@ -73,7 +67,4 @@ contract TestToken is BasedOFT {
         require(!tradingActivated, "Trading is already active");
         tokensaleContract = _tokensaleContract;
     }
-
-
-    
 }
